Modify clean command to remove RunDir
Change the builtin clean command so that it removes the RunDir, and
consider if clean should be run earlier. This will prevent the git
repo from growing too large with disconnected refs, as well as
ensure it can be cleaned even if the git repo gets into a broken
state.